Bring Windows integration test runtime down to less than half
This changelist parallelises the Windows integration test specs by modifying the integration test runner to accept arbitrary arguments to be passed to the `ginkgo` binary, allowing the integration test GitHub workflow to pass the `-p` and `--procs` arguments to parallelise running the test specs. This has brought the total runtime of Windows integration tests down to ~55 minutes from ~2 hours and 30 minutes.

The cluster version used in the tests has been updated to the latest EKS version 1.27, and as a result, one of the tests that uses the Docker runtime has been removed as it's no longer supported.