Fix checks for updated addon versions
When executing `eksctl get addons` an available update is not shown
if the only change to the version string is the `eksbuild` number, e.g.
`v1.0.0-eksbuild.2` is not shown as an update of `v1.0.0-eksbuild.1`.

The `findNewerVersions` func in `pkg/actions/get.go` had code to
explicitly ignore anything after the patch number. So `v1.1.0-eksbuild.1`
and `v1.1.0-eksbuild.4` were both converted to `1.1.0` and considered
equal. This fix removes the code that explicitly ignores the
`-eksbuild.x` portion of the version and allows the `semver` package
to compare the full version information.

Tests have also been updated to better match the version strings used
by AWS addons.